st-core/requirements.txt
bottle==0.12.17
paho-mqtt==1.5.0
SQLAlchemy==1.3.11
